Add tests

+++ file/file_test.go
@@ -16,15 +16,15 @@ func Test_ListFiles_ReturnsUpMigrationFiles_InCaseOfSuccess(t *testing.T) {
 	assert.Equal(t, 3, len(files))
 
 	assert.Equal(t, "1494538273_create-table-users.up.sql", files[0].Base)
-	assert.Equal(t, 1494538273, files[0].Version)
+	assert.Equal(t, int64(1494538273), files[0].Version)
 	assert.NotEmpty(t, files[0].SQL)
 
 	assert.Equal(t, "1494538317_add-phone-number-to-users.up.sql", files[1].Base)
-	assert.Equal(t, 1494538317, files[1].Version)
+	assert.Equal(t, int64(1494538317), files[1].Version)
 	assert.NotEmpty(t, files[1].SQL)
 
 	assert.Equal(t, "1494538407_replace-user-phone-with-email.up.sql", files[2].Base)
-	assert.Equal(t, 1494538407, files[2].Version)
+	assert.Equal(t, int64(1494538407), files[2].Version)
 	assert.NotEmpty(t, files[2].SQL)
 }
 
@@ -37,14 +37,14 @@ func Test_ListFiles_ReturnsDownMigrationFiles_InCaseOfSuccess(t *testing.T) {
 	assert.Equal(t, 3, len(files))
 
 	assert.Equal(t, "1494538407_replace-user-phone-with-email.down.sql", files[0].Base)
-	assert.Equal(t, 1494538407, files[0].Version)
+	assert.Equal(t, int64(1494538407), files[0].Version)
 	assert.NotEmpty(t, files[0].SQL)
 
 	assert.Equal(t, "1494538317_add-phone-number-to-users.down.sql", files[1].Base)
-	assert.Equal(t, 1494538317, files[1].Version)
+	assert.Equal(t, int64(1494538317), files[1].Version)
 	assert.NotEmpty(t, files[1].SQL)
 
 	assert.Equal(t, "1494538273_create-table-users.down.sql", files[2].Base)
-	assert.Equal(t, 1494538273, files[2].Version)
+	assert.Equal(t, int64(1494538273), files[2].Version)
 	assert.NotEmpty(t, files[2].SQL)
 }
